Introduce `monotonicity` parameter to the modules
This patch does introduces monotonicity parameter (non-decreasing or
non-increasing), which enforces the monotoicity of the control points
in the grids. This fine control allows to yield non-linear monotonic
dependencies. By default behaviour is preserved and monotonicity is
not enforced.

Also this patch updates formatting/linting: ruff version is updated,
black is dropped (they are doing the same and override output of each
other, which is very annoying). The preference is given to the ruff
since it was the first in the `.pre-commit-config.yaml`.